+++ package.json
@@ -24,7 +24,8 @@
   "dependencies": {
     "eslint": "^8.57.1",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.5.1"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^3.2.6",
@@ -39,6 +40,7 @@
     "@storybook/test": "^8.6.12",
     "@types/react": "^19.1.1",
     "@types/react-dom": "^19.1.2",
+    "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^8.30.1",
     "@typescript-eslint/parser": "^8.30.1",
     "@vitejs/plugin-react-swc": "^3.8.1",
