+++ src/components/CardDisplay/CardDisplay.tsx
@@ -1,4 +1,4 @@
-import CardExpirationDate from "../../types/CardExpirationDate";
+import CardExpirationDate from "src/types/CardExpirationDate";
 import { CardNumbers } from "../../types/CardNumbers";
 import CardBody from "./CardBody";
 import styles from "./cardDisplay.module.css";
