+++ src/components/Complete/CardComplete.tsx
@@ -4,12 +4,16 @@ import Button from "../common/Button/Button";
 import styles from "./CardComplete.module.css";
 
 const CardComplete = ({ newCard }) => {
+  if (Object.keys(newCard).length === 0) {
+    return <p>카드 정보가 없습니다</p>;
+  }
+
   return (
     <div className={styles.container}>
       <img src={cardComplete} alt="카드 등록 완료" />
       <p>{newCard.cardNumbers.fourthNumber}로 시작하는</p>
       <p>{newCard.cardCompany}가 등록되었어요.</p>
-      <Button text="확인" />
+      <Button link="/react-payments/" text="확인" />
     </div>
   );
 };
