+++ src/components/common/Button/Button.module.css
@@ -1,7 +1,13 @@
-.basic {
+.button {
     width: 375px;
     height: 52px;
     background-color: #000000;
-    font-size: 15px;
-    color: #ffffff
+    font-size: 15px;    
+    
+}
+
+.link {
+    text-decoration: none;
+    color: #ffffff;
+    
 }
