+++ src/components/common/Button/Button.tsx
@@ -1,9 +1,21 @@
+import { Link } from "react-router";
+
 import styles from "./Button.module.css";
 
-const Button = ({ text, onClick }: { text: string; onClick: () => void }) => {
+const Button = ({
+  text,
+  link,
+  onClick,
+}: {
+  text: string;
+  link?: string;
+  onClick?: () => void;
+}) => {
   return (
-    <button className={styles.basic} onClick={onClick}>
-      {text}
+    <button className={styles.button} onClick={onClick}>
+      <Link to={{ pathname: link }} className={styles.link}>
+        {text}
+      </Link>
     </button>
   );
 };
